aiPipeline-sdss2025/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.10
aiosignal==1.3.1
alabaster==1.0.0
annotated-types==0.6.0
anyio==4.6.2.post1
appnope==0.1.4
assemblyai==0.35.1
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.35.60
botocore==1.35.60
certifi==2024.8.30
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.15
distlib==0.3.9
distro==1.9.0
docutils==0.21.2
emoji==2.14.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
executing==2.0.1
ffmpeg==1.4
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.26.2
hyperframe==6.0.1
identify==2.6.2
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.1
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
jupyter_client==8.6.0
jupyter_core==5.7.1
langdetect==1.0.9
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.6
microsoft-kiota-abstractions==1.6.2
microsoft-kiota-authentication-azure==1.6.2
microsoft-kiota-http==1.6.2
microsoft-kiota-serialization-form==1.6.2
microsoft-kiota-serialization-json==1.6.2
microsoft-kiota-serialization-multipart==1.6.2
microsoft-kiota-serialization-text==1.6.2
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msgpack==1.1.0
msgraph-core==1.1.6
msgraph-sdk==1.12.0
multidict==6.1.0
mypy-extensions==1.0.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
olefile==0.47
openai==1.16.2
openpyxl==3.1.5
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
packaging==23.2
pandas==2.2.1
parso==0.8.3
pendulum==3.0.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.0
pluggy==1.5.0
pooch==1.8.2
portalocker==2.8.2
pre_commit==4.0.1
prompt-toolkit==3.0.43
propcache==0.2.0
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pypdf==5.1.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
RapidFuzz==3.10.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.0
sentence-transformers==3.3.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
Sphinx==8.1.3
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
std-uritemplate==2.0.1
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.6.0
time-machine==2.16.0
tokenizers==0.20.3
torch==2.5.1
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.1
transformers==4.46.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.0
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.17.1
zipp==3.21.0
